requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
coverage==5.5
Deprecated==1.2.13
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
IMAPClient==2.2.0
iniconfig==1.1.1
managesieve==0.7.1
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-asgi==0.36b0
opentelemetry-instrumentation-aws-lambda==0.36b0
opentelemetry-instrumentation-dbapi==0.36b0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-instrumentation-logging==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-sqlite3==0.36b0
opentelemetry-instrumentation-urllib==0.36b0
opentelemetry-instrumentation-urllib3==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
packaging==21.0
pluggy==1.0.0
prometheus-client==0.15.0
py==1.10.0
pycodestyle==2.7.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
sievelib==1.2.1
six==1.16.0
sqlparse==0.4.1
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.7
uvicorn==0.15.0
wrapt==1.14.1
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
coverage==5.5
Deprecated==1.2.13
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
IMAPClient==2.2.0
iniconfig==1.1.1
managesieve==0.7.1
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-asgi==0.36b0
opentelemetry-instrumentation-aws-lambda==0.36b0
opentelemetry-instrumentation-dbapi==0.36b0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-instrumentation-logging==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-sqlite3==0.36b0
opentelemetry-instrumentation-urllib==0.36b0
opentelemetry-instrumentation-urllib3==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
packaging==21.0
pluggy==1.0.0
prometheus-client==0.15.0
py==1.10.0
pycodestyle==2.7.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
sievelib==1.2.1
six==1.16.0
sqlparse==0.4.1
starlette==0.14.2
tnefparse==1.4.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.7
uvicorn==0.15.0
wrapt==1.14.1
